requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0